refactor: move AI logic to separate module
